+++ app/controllers/api/base_controller.rb
@@ -2,24 +2,6 @@
 
 module Api
   class BaseController < ::ApplicationController
-    def render_validation_errors(obj)
-      render400 messages: obj.errors.full_messages
-    end
-
-    def render400(messages:)
-      render_error(
-        code: 400,
-        title: 'Bad Request',
-        messages: messages
-      )
-    end
-
-    def render_error(code:, title:, messages:)
-      render json: {
-        code: code,
-        title: title,
-        messages: messages
-      }, status: code
-    end
+    include ExceptionRescuable
   end
 end
