+++ db/schemas/users.schema
@@ -1,6 +1,6 @@
-create_table "users", force: :cascade, options: "COLLATE=utf8_bin" do |t|
-  t.string "qrcode", null: false
-  t.string "name", null: false
+create_table :users, force: :cascade, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_bin" do |t|
+  t.string   :qrcode,   null: false
+  t.string   :name,     null: false
 end
 
 add_index :users, :qrcode, unique: true
